\name{ROCS-package}
\alias{ROCS-package}
\alias{ROCS}
\docType{package}
\title{
Receiver Operating Characteristics Surface for class-skewed data
}
\description{
The package constructs Receiver Operating Characteristics Surface (ROCS) using the false positive rate (FPR), true positive rate (TPR) and the false discovery rate (FDR). It computes the volume under the surface (VUS) and FDR-controlled area under the curve (FCAUC) to gauge the separability between two classes while taking into account the proportion of false discoveries.
}
\details{
\tabular{ll}{
Package: \tab ROCS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-12\cr
License: \tab GPL 2.0 or later\cr
}
Major functions are rocs.x and fcauc.x. 
}
\author{
Tianwei Yu

Maintainer: Tianwei Yu <tyu8@emory.edu>
}
\keyword{ package }
